Add test for initializeWebSocket
